move file reading code to a separate... file
